# api/api.yaml
openapi: 3.1.0
info:
  title: Polygon ID
  description: |
    Documentation for the Backend Verifier API
  version: "1"
servers:
  - description: Local
    url: http://localhost:3010
...
paths:
...
  /sign-in:
    post:
      summary: Sign in
      operationId: Sign in
      description: |
        Sign in with a verification request
      tags:
        - Verifier
      requestBody:
        content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignInRequest'
              examples:
                OffChain:
                  value:
                    {
                      "chainID": "80001",
                      "skipClaimRevocationCheck": false,
                      "scope": [
                        {
                          "circuitID": "credentialAtomicQuerySigV2",
                          "query": {
                            "context": "https://raw.githubusercontent.com/iden3/claim-schema-vocab/main/schemas/json-ld/kyc-v3.json-ld",
                            "allowedIssuers": [ "*" ],
                            "type": "KYCAgeCredential",
                            "credentialSubject": {
                              "birthday": {
                                "$eq": 19960424
                              }
                            }
                          }
                        }
                      ]
                    }
                OnChain:
                  value:
                    {
                      "chainID": "80001",
                      "skipClaimRevocationCheck": false,
                      "scope": [
                        {
                          "circuitID": "credentialAtomicQuerySigV2",
                          "query": {
                            "context": "https://raw.githubusercontent.com/iden3/claim-schema-vocab/main/schemas/json-ld/kyc-v3.json-ld",
                            "allowedIssuers": [ "*" ],
                            "type": "KYCAgeCredential",
                            "credentialSubject": {
                              "birthday": {
                                "$eq": 19960424
                              }
                            }
                          }
                        }
                      ],
                      "transactionData": {
                        "contractAddress": "0xE826f870852D7eeeB79B2C030298f9B5DAA8C8a3",
                        "methodID": "b68967e2",
                        "chainID": 80001,
                        "network": "polygon-mumbai"
                      }
                    }
      responses:
        '200':
          description: Authorization Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingInResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /qr-store:
    post:
      summary: Store QRCode
      operationId: QRStore
      tags:
        - Verifier
      requestBody:
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/QRStoreRequest'
      responses:
        '200':
          description: QR Indirection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRStoreResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get QRCode from store
      operationId: GetQRCodeFromStore
      tags:
        - Verifier
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: QR Code Indirection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCode'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  /callback:
    post:
      summary: Callback
      operationId: Callback
      tags:
        - Verifier
      parameters:
        - $ref: '#/components/parameters/sessionID'
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: jwz-token
              description: JWZ token
      responses:
        '200':
          description: Callback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackResponse'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
components:
  schemas:
    CallbackResponse:
      type: object
    QRStoreResponse:
      type: string
      example: iden3comm://?request_uri=https%3A%2F%2Fissuer-demo.polygonid.me%2Fapi%2Fqr-store%3Fid%3Df780a169-8959-4380-9461-f7200e2ed3f4
    QRStoreRequest:
      $ref: '#/components/schemas/QRCode'
      example: ' {
       "id":"6dc645a6-2be3-4099-a645-20784ee53cd0",
       "typ":"application/iden3comm-plain-json",
       "type":"https://iden3-communication.io/authorization/1.0/request",
       "thid":"6dc645a6-2be3-4099-a645-20784ee53cd0",
       "body":{
          "callbackUrl":"https://verifier-backend/callback?sessionId=6dc645a6-2be3-4099-a645-20784ee53cd0",
          "reason":"test flow",
          "scope":[
             {
                "id":1,
                "circuitId":"credentialAtomicQuerySigV2",
                "query":{
                   "allowedIssuers":[
                      "*"
                   ],
                   "context":"https://raw.githubusercontent.com/iden3/claim-schema-vocab/main/schemas/json-ld/kyc-v4.jsonld",
                   "credentialSubject":{
                      "birthday":{
                         "$lt":20000101
                      }
                   },
                   "skipClaimRevocationCheck":true,
                   "type":"KYCAgeCredential"
                }
             }
          ]
       },
       "from":"did:polygonid:polygon:mumbai:2qH7TstpRRJHXNN4o49Fu9H2Qismku8hQeUxDVrjqT"
      } '
...
    SingInResponse:
      type: object
      required:
          - sessionID
          - qrCode
      properties:
          sessionID:
            $ref: '#/components/schemas/UUID'
          qrCode:
            $ref: '#/components/schemas/QRCode'
    QRCode:
      type: object
      required:
          - id
          - typ
          - type
          - thid
          - body
          - from
      properties:
        id:
          type: string
          example: '6dc645a6-2be3-4099-a645-20784ee53cd0'
        typ:
          type: string
          example: 'application/iden3comm-plain-json'
        type:
          type: string
          example: 'https://iden3-communication.io/authorization/1.0/request'
        thid:
          type: string
          example: '6dc645a6-2be3-4099-a645-20784ee53cd0'
        body:
          $ref: '#/components/schemas/Body'
        from:
          type: string
          example: 'did:polygonid:polygon:mumbai:2qH7TstpRRJHXNN4o49Fu9H2Qismku8hQeUxDVrjqT'
        to:
          type: string
          example: 'did:polygonid:polygon:mumbai:2qEATqfECVbCBzq9EhJpPSiv1xtJRpbMBKDaNM68Ci'
    Body:
      type: object
      required:
        - reason
        - scope
      properties:
        transaction_data:
          $ref: '#/components/schemas/TransactionDataResponse'
        callbackUrl:
          type: string
          example: 'https://verifier-backend/callback?sessionId=6dc645a6-2be3-4099-a645-20784ee53cd0'
        reason:
          type: string
          example: 'test flow'
        scope:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
    Scope:
      type: object
      required:
        - id
        - circuitId
        - query
      properties:
        id:
          type: integer
          example: 1
        circuitId:
          type: string
          example: 'credentialAtomicQuerySigV2'
        query:
          $ref: '#/components/schemas/Query'
        params:
          $ref: '#/components/schemas/ScopeParams'
      example:
        {
          "circuitID": "credentialAtomicQuerySigV2",
          "query": {
            "context": "https://raw.githubusercontent.com/iden3/claim-schema-vocab/main/schemas/json-ld/kyc-v3.json-ld",
            "allowedIssuers": [ "*" ],
            "type": "KYCAgeCredential",
            "credentialSubject": {
              "birthday": {
                "$eq": 19960424
              }
            }
          }
        }
    Query:
      type: object
      example:
        {
           "context": "https://raw.githubusercontent.com/iden3/claim-schema-vocab/main/schemas/json-ld/kyc-v3.json-ld",
           "allowedIssuers": ["*"],
           "type": "KYCAgeCredential",
           "credentialSubject": {
               "birthday": {
                   "$lt": 20200424
               }
           }
        }
    SignInRequest:
      type: object
      required:
        - scope
      properties:
        chainID:
          type: string
          description: |
            Only required when using off-chain verification
            `80001`: `mumbai`
            `137` : `mainnet`
          example: '80001'
        reason:
          type: string
          example: 'test flow'
        to:
          type: string
          example: null
        scope:
          type: array
          items:
            $ref: '#/components/schemas/ScopeRequest'
        transactionData:
          $ref : '#/components/schemas/TransactionData'
    ScopeRequest:
      type: object
      required:
        - circuitId
        - query
      properties:
        circuitId:
          type: string
          example: 'credentialAtomicQuerySigV2'
        query:
          $ref: '#/components/schemas/Query'
        params:
          $ref: '#/components/schemas/ScopeParams'
    ScopeParams:
      type: object
      example:
        {
          "nullifierSessionId": "123443290439234342342423423423423"
        }
    TransactionDataResponse:
      type: object
      description: |
        Only required when using on-chain verification
      required:
        - contract_address
        - method_id
        - chain_id
        - network
      properties:
        contract_address:
          type: string
          example: '0xE826f870852D7eeeB79B2C030298f9B5DAA8C8a3'
        method_id:
          type: string
          example: 'b68967e2'
        chain_id:
          type: integer
          example: 80001
        network:
          type: string
          example: 'polygon-mumbai'
    TransactionData:
      type: object
      description: |
        Only required when using on-chain verification
      required:
        - contractAddress
        - methodID
        - chainID
        - network
      properties:
        contractAddress:
          type: string
          example: '0xE826f870852D7eeeB79B2C030298f9B5DAA8C8a3'
        methodID:
          type: string
          example: 'b68967e2'
        chainID:
          type: integer
          example: 80001
        network:
          type: string
          example: 'polygon-mumbai'
    UUID:
      type: string
      x-go-type: uuid.UUID
      x-go-type-import:
        name: uuid
        path: github.com/google/uuid
      example: 8edd8112-c415-11ed-b036-debe37e1cbd6
    GenericErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Something happen'
  parameters:
    sessionID:
      name: sessionID
      in: query
      required: true
      description: |
        ID e.g: 89d298fa-15a6-4a1d-ab13-d1069467eedd
      schema:
        type: string
        x-go-type: uuid.UUID
        x-go-type-import:
          name: uuid
          path: github.com/google/uuid
    id:
      name: id
      in: query
      required: true
      description: |
        ID e.g: 89d298fa-15a6-4a1d-ab13-d1069467eedd
      schema:
        type: string
        x-go-type: uuid.UUID
        x-go-type-import:
          name: uuid
          path: github.com/google/uuid
  responses:
    '400':
      description: 'Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
    '404':
      description: 'Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
    '500':
      description: 'Internal Server error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'